Bug Fixes
FitMeanAPMultipleSets takes 3 and 7 as the initial parameters for
TimeStart and TimeEnd, where as version 2 takes 4 and 8 respectively.
This was initially to test something about the data, but both versions
are no longer needed, as they do virtually the same thing.
Generate Figures is now added, which take an input of (nuclear cycle,
data type, AP range)
nuclear cycle ranges from 12-14, data type is the data type name that
matches what is found on the DataStatus.xlsx sheet, and AP axis is the
range of AP you want to view the histograms from. Can be a single bound
(i.e. 0.6) or two (0.1, 0.7).
DataStatus.xlsx file is needed in the same location as
MovieDatabase.xlsx