Created a template for a standard script documentation header
While looking through our current code, I realized it would be very
helpful to have a standard documentation header at the top of each
script or function. Right now, everyone's scripts are documented in
different ways, making it difficult to systematically compile
information about all our code.

I have added a script, "documentationTemplate", to the master
mRNADynamics branch, which contains a template/example of a standard
documentation header that everyone should put on their code. Please make
an effort to both include the standard header in any new code you add
and go back into any code you've written and put a proper header on it.

The most important detail to note is that the header must be in a single
comment block (i.e. if you want to add blank lines in between sections,
you need to have a percent sign '%' at the start to make a commented
line, as shown in the template). This is because we will likely be using
MATLAB's own "help" and "doc" functions to compile documentation, and
they pull everything in the first uninterrupted comment block. If you're
unsure about this, feel free to ask me.

I know this might take a bit of work, but this will make our lives
easier in the future. If you have any questions about the header or any
suggestions for additions/changes you think would be helpful, please let
me know!