Introduced unit testing
I added a folder in the repo called "testClasses". This folder will contain unit tests for functions in the repo. The idea is they'll all test the sample data set (although other data sets should work, mostly). I only added a test class for segmentSpots. This can be used as a model for future classses.