Added functionality for new "lattice" ExperimentType
Added ability to read in "lattice" as an ExperimentType.

Treats "lattice" the same as "inputoutput" when naming TIF files and looking for them later, where the His and MCP channels are not fixed and can be Channel1 or Channel2.

Also allows the code to run even if you haven't produced an Ellipses.mat due to a bad Histone channel, with a warning that Ellipses.mat doesn't exist and the code is proceeding without information from a Histone channel.