Not reassigning dropbox folder
The reassignment of the dropbox folder to the default dropbox folder was commented out from the try catch section that tries to load the experiment data from movie database.

The assumption made is that the previous chunk of code correctly found the folder of the datastatus of interest and that folder is also where the CompiledParticles.mat of the movies of interests are as well.