ExportToAWS now only copies the "stacks" folder in PreProcessedData
This should expedite the syncing process on Dropbox, as only the z-stacks in the stacks folder are used (the indiv frame files in PreProcessed aren't used by filterMovie).
Also added command window display statements so it's easier to keep track of progress when your running multiple prefixes in a row.