Massive overhaul
Changes of note:

-1) The pipeline uses 3D tif stacks for everything now. There are many upsides to this, but a couple issues to note- some loading steps may be slower. Loading speed is strongly affected by having more parallel workers, though, so that is the recommended approach. After initial loading, everything is quite fast until the Prefix is changed or Matlab is closed.

.lif files no longer need to be exported from LASX. the pipeline has its own export script now that runs automatically.

Entire movies kept in memory during the duration of use (while Matlab is open or until the Prefix is changed). For the most part, that means faster clicking but slower initial loading. The initial load is speed is parallelizable. Additionally, note that you may get strange errors sometimes due to the fact that movies are hidden in the background and not obviously in the workspace. This means "clear all" is now an important troubleshooting step if you run into unexplained problems.

Usual segmentation algorithm in TrackNuclei changed to something more robust. Works far better in difficult cases, and seems to work just as well with typical cases (e.g. good histone). Note that the nuclei are actually stored as full-blown ellipses now, and not circles. I found that this somehow improves instance segmentation. It's nice for other reasons too like nuclear masking, since the ellipses are usually snuggly wrapped around the real nuclear boundaries.

chooseAnaphaseFrames is the preferred method for documenting mitosis timings instead of moviedatabase. Both can still be used.

determineThreshold is the preferred method of determining the DoG or probability threshold for spot segmentation. This is called automatically by segmentspots.

Nuclear masking is turned on by default in segmentspots. typically, that means tracknuclei should be run first. If no Ellipses.mat is detected, that step will be skipped.

CheckNucleiSegmentation has a nice manual tracking and segmentation feature using "c" or "v" to track from the past or future.

CheckParticleTracking has a functional multiview option.

liveExperiment is a class created to keep track of an entire experiment. This is the interface I use for getting folder information about a project, for getting moviedatabase information, and for loading movies and data structures.