Separating the trace figure from the movie figure
There have been a couple of issues in the figures shown by the CheckPArticleTracking. The issues were that the TraceFig and OverlayFig sometimes overlap, and also the field of view of OverlayFig is not shown as a whole in case of 1024x256 format, for example. This is due to the MATLAB assinging the axis size arbitrarily. So, what I did here is
1) Separating the TraceFig from the OverlayFig
2) Position of the TraceFig in the screen. This can be adjusted later. I think it's something one can adjust using options as the movie format is different for different projects.