Readded parfor loop in segmentTranscriptionalLoci
This had been commented out because segmentTranscriptionalLoci was throwing an elusive, currenlty unsolved bug related to the shouldMaskNuclei user option.

**Currently, you can either use the parfor loop without the shouldMaskNuclei option (will run at normal speeds) or use the shouldMaskNuclei option but run segmentation with a normal for loop (it will take MUCH longer to run)***

The code will now throw a preemptive error when a user passes the 'shouldMaskNuclei' option. The error includes the following instructions on how to make the code run with that option:

"(1) In segmentTranscriptionalLoci.m, change the parfor loop to a for loop (just delete the letters ''par'') at approx. line 127.
(2) In segmentTranscriptionalLoci.m, find the two 'if shouldMaskNuclei' statements, at approx. lines 95 and 198. Uncomment both these if statements in their entirety.
(3) In determineSegmentSpotsOptions.m, comment out this error message at approx. line 82.
(4) DO NOT push the changes in 1-3 to the master branch. If you need them, please stash on a separate branch.
If you are confused or need help implementing the steps above, please contact Meghan Turner (meghan_turner@berkeley.edu)"