Update to preemptive error message in segmentTranscriptionalLoci
***Correction to the previous commit***

The nuclearMask option has been commented out because segmentTranscriptionalLoci was throwing an elusive, currenlty unsolved bug whereby it would always try to implement the 'nuclearMask' option, even if the user set it to false.

UPDATED: The parfor loop can be used either with or without the 'nuclearMask' option. The code currently has all statements related to 'nuclearMask' commented out so that you can use the parfor loop without the nuclearMask option. If you need to use the nuclearMask option, follow the steps below to uncomment to necessary lines of code:

The code will now throw a preemptive error when a user passes the 'shouldMaskNuclei' option. The error includes the following instructions on how to make the code run with that option:

"(1) In segmentTranscriptionalLoci.m, find the two 'if shouldMaskNuclei' statements, at approx. lines 95 and 198. Uncomment both these if statements in their entirety.
(2) In determineSegmentSpotsOptions.m, comment out this error message at approx. line 82.
(3) DO NOT push the changes in 1-2 to the master branch. If you need them, please stash on a separate branch.
If you are confused or need help implementing the steps above, please contact Meghan Turner (meghan_turner@berkeley.edu)"