Adds "ComputeAltFluo" flag to CompileNuclearProtein
Adds "ComputeAltFluo" flag for CompileNuclearProtein, which computes nuclear fluorescence using the median fluorescence value of the 10 z slices around the maximum fluorescence slice. This was added to prevent errors when nuclei move within the z stack such that the center of the stack does not correspond to the center of the nucleus.