Hacky fix for bug that erased nuclear Fluo data
When you ran CheckParticleTracking with a Nucleus associated with the
current particle, it lead to the erasure of the integrated nuclear fluorescence trace that's stored in schnitzcells for an input experiment.
I don't know what case it was supposed to catch, so I just commented it out rather than trying to correct it. Hasn't caused me issues yet ...